+++ sdg/src/main/java/com/shopl/sdg/component/text_input/SDGInput.kt
@@ -3,21 +3,17 @@ package com.shopl.sdg.component.text_input
 import android.annotation.SuppressLint
 import android.graphics.Rect
 import android.view.ViewTreeObserver
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
-import androidx.compose.foundation.gestures.animateScrollBy
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.BasicTextField
 import androidx.compose.foundation.text.KeyboardOptions
@@ -30,8 +26,6 @@ import androidx.compose.runtime.State
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusRequester
@@ -40,12 +34,10 @@ import androidx.compose.ui.focus.focusRequester
 import androidx.compose.ui.focus.onFocusChanged
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.SolidColor
-import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.platform.LocalView
 import androidx.compose.ui.platform.TextToolbar
 import androidx.compose.ui.platform.TextToolbarStatus
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.TextRange
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontFamily
@@ -56,11 +48,9 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.shopl.sdg_common.enums.Keyboard
-import com.shopl.sdg_common.ext.clickable
 import com.shopl.sdg_common.foundation.SDGColor
 import com.shopl.sdg_common.ui.components.IOText
 import com.shopl.sdg_common.ui.components.TypefaceConfig
-import com.shopl.sdg_resource.R
 import java.text.DecimalFormat
 
 
@@ -72,125 +62,7 @@ import java.text.DecimalFormat
  *      SDGUnderlineInput()
  *  }
  */
-@Composable
-fun SDGUnderlineInput(
-    input: String?,
-    hint: String,
-    inputState: InputState,
-    backgroundColor: Color = SDGColor.Neutral200,
-    marginValues: PaddingValues = PaddingValues(),
-    onInputChange: (String) -> Unit,
-) {
-    val coroutineScope = rememberCoroutineScope()
-    val scrollState = rememberScrollState()
-    val focusRequester by remember { mutableStateOf(FocusRequester()) }
-    val focusManager = LocalFocusManager.current
 
-    val isKeyboardOpen by keyboardAsState()
-
-    val lineColor = when (inputState) {
-        InputState.Enable -> backgroundColor
-        InputState.Disable -> backgroundColor
-        is InputState.Error -> SDGColor.Red300
-    }
-
-    val textColor = when (inputState) {
-        InputState.Disable -> SDGColor.Neutral300
-        else -> SDGColor.Neutral700
-    }
-
-    val inputModifier = Modifier
-        .padding(marginValues)
-        .fillMaxWidth()
-        .focusRequester(focusRequester = focusRequester)
-        .onFocusChanged {
-            //Log.e("SDGInput", "Focus : ${it.isFocused}")
-        }
-
-    val textStyle = TextStyle(
-        color = textColor,
-        fontSize = 18.sp,
-        fontFamily = FontFamily(typeface = TypefaceConfig.normal),
-        letterSpacing = 0.sp,
-        lineHeight = 22.sp,
-    )
-
-    var maxHeight by remember { mutableStateOf(-1) }
-    val lineHeight = with(LocalDensity.current) { 22.sp.toPx() }
-
-    LaunchedEffect(key1 = input) {
-        if (maxHeight < scrollState.maxValue) {
-            scrollState.animateScrollBy(lineHeight)
-            maxHeight = scrollState.maxValue
-        }
-    }
-
-    LaunchedEffect(key1 = isKeyboardOpen) {
-        //Log.e("SDGInput", "SideEffect isKeyboardOpen : $isKeyboardOpen")
-        if (isKeyboardOpen == Keyboard.Closed) {
-            focusManager.clearFocus()
-        }
-    }
-
-    BasicTextField(
-        modifier = inputModifier,
-        value = input ?: "",
-        onValueChange = {
-            onInputChange(it)
-        },
-        enabled = inputState != InputState.Disable,
-        textStyle = textStyle,
-        singleLine = true,
-        cursorBrush = SolidColor(SDGColor.Neutral700),
-        decorationBox = { textField ->
-            Column(
-                Modifier
-                    .fillMaxWidth()
-                    .padding(
-                        top = 8.dp,
-                        bottom = 8.dp
-                    )
-            ) {
-                Row(
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Box(Modifier.weight(1F)) {
-                        textField()
-                        if (input.isNullOrEmpty()) {
-                            IOText(
-                                text = hint,
-                                textColor = SDGColor.Neutral300,
-                                fontSize = 18.sp,
-                                lineHeight = 22.sp,
-                            )
-                        }
-                    }
-
-                    if (!input.isNullOrEmpty() && isKeyboardOpen == Keyboard.Opened) {
-                        Image(
-                            modifier = Modifier
-                                .padding(start = 10.dp)
-                                .clickable {
-                                    onInputChange.invoke("")
-                                },
-                            painter = painterResource(
-                                id = R.drawable.ic_input_delete
-                            ),
-                            contentDescription = null
-                        )
-                    }
-                }
-                Spacer(
-                    modifier = Modifier
-                        .padding(top = 4.dp)
-                        .fillMaxWidth()
-                        .height(1.dp)
-                        .background(color = lineColor)
-                )
-            }
-        },
-    )
-}
 
 @SuppressLint("NewApi")
 @Composable
@@ -653,16 +525,6 @@ private fun PrevInput(
                 .fillMaxSize()
                 .background(color = SDGColor.Neutral0)
         ) {
-
-
-            SDGUnderlineInput(
-                marginValues = PaddingValues(20.dp),
-                //input = "1234123412341234123412341234123412341234123412341234123412341234",
-                input = "12341234",
-                hint = "asdfasdf",
-                inputState = InputState.Enable,
-                onInputChange = {}
-            )
 
             SDGSimpleTextInput(
                 type = SDGSimpleInputType.BASIC,
